Reorganize compiler
