--- tutara-compiler-llvm/src/compiler/mod.rs
@@ -1,9 +1,0 @@
-#![allow(clippy::module_inception)]
-mod compiler;
-pub use compiler::Compiler;
-
-mod evaluator;
-pub use evaluator::Evaluator;
-
-mod scope;
-pub use scope::Scope;
