+++ tutara-compiler-llvm/src/lib.rs
@@ -1,2 +1,14 @@
-pub mod compiler;
+mod compiler;
 pub use compiler::*;
+
+mod evaluator;
+pub use evaluator::*;
+
+mod scope;
+pub use scope::*;
+
+mod operation;
+pub use operation::*;
+
+mod compilation;
+pub use compilation::*;
